feat(entity): 统一集群ID字段类型为Long，提升数据一致性
- 将多个实体类中的集群ID字段类型从Integer更改为Long，以确保数据模型的一致性与兼容性。
- 涉及的实体类包括AlertGroupEntity、AuthTokenEntity、AutoScaleTaskEntity、ClusterAlertExpressionEntity等，确保所有相关逻辑均能正确处理集群ID。

这些改动显著增强了数据模型的健壮性与一致性，确保了整体设计的一致性。

+++ datasophon-dao/src/main/java/com/datasophon/dao/entity/ClusterServiceInstanceConfigEntity.java
@@ -65,7 +65,7 @@ public class ClusterServiceInstanceConfigEntity implements Serializable {
     /**
      *
      */
-    private Integer clusterId;
+    private Long clusterId;
 
     private String configFileJson;
